+++ src/main/java/study/Application.java
@@ -8,7 +8,7 @@ public class Application {
 	public static void main(String[] args) {
 		String expression = InputView.inputExpression();
 		String[] values = expression.split(" ");
-		int result = (int)Calculator.calculate(values);
+		int result = Calculator.calculate(values);
 		OutputView.printResult(result);
 	}
 }
